feat(table): complete refactor to make table components independent - Created AppTable component to centralize table management - Imported Table, PaginationTable, and SearchBarTable as child components - Passed data to Table via props to ensure full independence - Removed old test Table file to avoid confusion - Renamed and reorganized files and components for better readability and consistency
