+++ src-react/App.jsx
@@ -1,9 +1,7 @@
 import { useEffect, useState } from "react";
 import axios from "axios";
-import Label from "./ui/Label";
 import { LogIn, Mail, LockKeyhole, MoveRight } from "lucide-react";
-import LoginForm from "./ui/LoginForm";
-
+import Table from "../src-react/components/ui/Table/AppTable"
 export default function App() {
     const [showAlert, setShowAlert] = useState(true);
 
@@ -26,7 +24,8 @@ export default function App() {
 
     return (
         <div>
-            <LoginForm initialStep="login" />
+            {/* <LoginForm initialStep="login" /> */}
+            <Table></Table>
         </div>
     );
 }
